+++ page_objects/cart_preview/cart_preview.py
@@ -1,7 +1,7 @@
 from page_objects.abstract.abs_page_object import AbsPageObject
 from page_objects.abstract.web_elements.button import Button
 from page_objects.abstract.web_elements.text import Text
-from page_objects.cart_preview_locators import CartPreviewLocators as Locators
+from page_objects.cart_preview.cart_preview_locators import CartPreviewLocators as Locators
 
 class CartPreview(AbsPageObject):
