+++ page_objects/product_list/product_list.py
@@ -3,6 +3,7 @@ from selenium_datatable import DataTable, Column
 
 
 class ProductList(DataTable):
+    """Appears both on category page and search page"""
 
     rows_locator = (By.CSS_SELECTOR, ".product_list > li")
